restructure python files + remove pycache
