+++ opengs_maptool/logic/utils.py
@@ -1,4 +1,4 @@
-import config
+import opengs_maptool.config as config
 import numpy as np
 from PIL import Image
 from scipy.spatial import cKDTree
