+++ opengs_maptool/main.py
@@ -1,6 +1,6 @@
 import sys
 from PyQt6.QtWidgets import QApplication
-from ui.main_window import MainWindow
+from opengs_maptool.ui.main_window import MainWindow
 
 
 if __name__ == "__main__":
