+++ opengs_maptool/ui/image_display.py
